requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile requirements.in
#
defusedxml==0.7.1
    # via fpdf2
fonttools==4.51.0
    # via fpdf2
fpdf2==2.7.8
    # via -r requirements.in
pillow==10.3.0
    # via
    #   -r requirements.in
    #   fpdf2
tomli==2.0.1
    # via -r requirements.in
